Resist temptation to request same certificate in a loop
When requesting a certificate, kube-lego drops unreachable domains from
the request. While this allows partial certificates to be provisioned,
as far as kube-lego is concerned the certificate remains unfulfilled,
and kube-lego continues in its attempts to deliver.

In the face of domains that remain unreachable for extended periods of
time (e.g., permanently), kube-lego would repeatedly request a
partial certificate, eventually triggering Let's Encrypt rate-limit for
duplicate certificates.

With this change, kube-lego keeps track of missing domains in an
existing certificate, and will only request a new certificate if it
includes at least one of the missing domains.

This prevents certificate request loops where one or more domains are
unreachable permanently (or for an extended period of time). The loop
condition remains if the set of unreachable domains is not stable.

This also prevents kube-lego from requesting a certificate for a
narrower set of domains, than currently supported by an existing
certificate.

Logic in newCertNeeded() was tweaked so that expired certificates are
checked for before checking for missing domains, to allow for partial
certificates to be renewed, and not skipped.

Fixes: https://github.com/jetstack/kube-lego/issues/330
Signed-off-by: João Abecasis <joao@schibsted.com>